Add admin function to lock user project, to be done while project error is being investigated.

+++ panel.manageDataset.php
@@ -146,9 +146,11 @@
 		$projectNameString  = trim($projectNameString);
 		echo "<span id='p_label_".$key."' style='color:#".$labelRgbColor.";'>\n\t\t\t\t";
 		echo "<font size='2'>".($key+1).".";
-		echo "<button id='project_delete_".$key."' type='button' onclick=\"parent.deleteProjectConfirmation('".$project."','".$key."');\">Delete</button>";
-		if (!file_exists("users/".$user."/projects/".$project."/minimized.txt") && file_exists("users/".$user."/projects/".$project."/complete.txt")) {
-			echo "<button id='project_minimize_".$key."' type='button' onclick=\"parent.minimizeProjectConfirmation('".$project."','".$key."');\">Minimize</button>";
+		if (!file_exists("users/".$user."/projects/".$project."/locked.txt")) {
+			echo "<button id='project_delete_".$key."' type='button' onclick=\"parent.deleteProjectConfirmation('".$project."','".$key."');\">Delete</button>";
+			if (!file_exists("users/".$user."/projects/".$project."/minimized.txt") && file_exists("users/".$user."/projects/".$project."/complete.txt")) {
+				echo "<button id='project_minimize_".$key."' type='button' onclick=\"parent.minimizeProjectConfirmation('".$project."','".$key."');\">Minimize</button>";
+			}
 		}
 		echo $projectNameString;
 
@@ -188,16 +190,21 @@
 
 		if ($frameContainerIx == "1") {
 			// define update dataset button, which passes key value to update project page in iframe of main page.
-			if (!file_exists("users/".$user."/projects/".$project."/minimized.txt") && file_exists("users/".$user."/projects/".$project."/complete.txt")) {
-				echo "<button id='project_update_".$key."' type='button' onclick='";
-				echo "parent.document.getElementById(\"Hidden_UpdateDataset_Frame\").contentWindow.location.href = \"project.update_window.php?key=".$key."\";\n";
-				echo "parent.show_hidden(\"Hidden_UpdateDataset\"); ";
-				echo "getElementById(\"project_update_".$key."\").style.display = \"none\";";
-				echo "getElementById(\"project_minimize_".$key."\").style.display = \"none\";";
-				echo "'>Update Dataset</button><br>";
+			if (!file_exists("users/".$user."/projects/".$project."/locked.txt")) {
+				if (!file_exists("users/".$user."/projects/".$project."/minimized.txt") && file_exists("users/".$user."/projects/".$project."/complete.txt")) {
+					echo "<button id='project_update_".$key."' type='button' onclick='";
+					echo "parent.document.getElementById(\"Hidden_UpdateDataset_Frame\").contentWindow.location.href = \"project.update_window.php?key=".$key."\";\n";
+					echo "parent.show_hidden(\"Hidden_UpdateDataset\"); ";
+					echo "getElementById(\"project_update_".$key."\").style.display = \"none\";";
+					echo "getElementById(\"project_minimize_".$key."\").style.display = \"none\";";
+					echo "'>Update Dataset</button><br>";
+				}
 			}
 		}
 
+		if (file_exists("users/".$user."/projects/".$project."/locked.txt")) {
+			echo "<font size='2' color='red'>[Project locked for admin review.]</font>";
+		}
 
 		echo "\t\t\t\t";
 		echo "<div id='frameContainer.p".$frameContainerIx."_".$key."'></div>\n\n\t\t\t\t";
